+++ src/filters/transform/vsfilter/plugins.cpp
@@ -295,7 +295,7 @@ public:
 
 		m_hThread = 0;
 
-		for(int i = 1; i < handles.GetCount(); i++)
+		for(size_t i = 1; i < handles.GetCount(); i++)
 			FindCloseChangeNotification(handles[i]);
 
 		return 0;
